vizjson 2 specs also for 3
